+++ crue10/modele.py
@@ -51,6 +51,8 @@ class Modele(EnsembleFichiersXML):
         :type files: dict(str)
         :param metadata: dictionnaire avec les méta-données
         :type metadata: dict(str)
+        :param version_grammaire: version de la grammaire
+        :type version_grammaire: str
         """
         check_preffix(nom_modele, 'Mo_')
         self.id = nom_modele