+++ crue10/utils/crueconfigmetier.py
@@ -4,6 +4,7 @@ import xml.etree.ElementTree as ET
 from crue10.utils import DATA_FOLDER_ABSPATH, ExceptionCrue10, PREFIX
 
 
+#: Chemin vers le fichier CCM
 CCM_FILE = os.path.join(DATA_FOLDER_ABSPATH, 'CrueConfigMetier.xml')